setting column increments all set

+++ database/migrations/2019_04_08_220852_create_properties_table.php
@@ -16,7 +16,7 @@ class CreatePropertiesTable extends Migration
         Schema::create('properties', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('user_id');
-            $table->unsignedInteger('bookdata_id');
+            $table->unsignedBigInteger('bookdata_id');
             $table->integer('number');
             $table->date('getdate')->nullable();
             $table->text('freememo')->nullable();
